added button for closed contracts if no rating provided
